requirements_dev.txt
# includes additional requirements needed for development of this project

-r requirements.txt

PyPDF2
matplotlib
seaborn
tabulate
psutil
holidays

# loading xlsx file with pandas
openpyxl

# LPG example
utspclient

# City simulation example
geopandas
contextily

pytest
sphinx
sphinx-autodoc-typehints

# type annotations
pandas-stubs
types-setuptools
types-tabulate
types-tqdm